automation/requirements.txt
# automation/requirements.txt
# Python dependencies for GenAI Pipeline Automation

# Google Drive & OAuth
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0

# Utilities
pydantic>=2.0.0
networkx>=3.0
tqdm>=4.66.0
jinja2>=3.1.0

# Optional: For docstring insertion
libcst>=1.4.0

# Optional: For advanced logging
python-json-logger>=2.0.0
